+++ src/lib/registry/charts/layerchart-area-chart/area-chart.svelte
@@ -101,7 +101,6 @@
 	let selectedDataKey = $state<string | null>(untrack(() => defaultSelectedDataKey));
 
 	const loading = new LoadingDataState({
-		isLoading: () => isLoading,
 		loadingPoints: () => loadingPoints ?? 14
 	});
 
@@ -315,7 +314,7 @@
 			{@render children()}
 			<TooltipCursor />
 			{#if isLoading}
-				<LoadingArea {chartId} {curveType} onShimmerExit={loading.onShimmerExit} />
+				<LoadingArea {chartId} {curveType} />
 			{/if}
 		</Svg>
 		<TooltipRender />